requirements.txt
Django==5.0.0
django-extensions==3.1.1
djangorestframework==3.14
ipython==7.21.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6
django-cities-light==3.10.0
django-filter==23.5
requests==2.31.0
